Read the time from the DS1307.
We set the DS1307 to a fixed time at startup, then read time
periodically from it afterwards, watching the time increment.

There are lots of error conditions that are not handled correctly,
or not handled at all.  The code is very much in a proof of concept
state, but I want to move it to master to show progress, and to use
this as a base for improvement.

Signed-off-by: Russell Steicke <russells@adelie.cx>