Read simple commands from the serial port.
The only commands working are TRON and TROFF, for changing the serial
output (tracing) state.

There are now two sets of serial output functions.  Those that start
with "serial_trace" are intended for verbose debugging output.  The
old "serial_send" functions will always produce output, and are
intended for startup messages and for output that should always be
printed, eg output from commands.

The macros in serial.h have also changed.  ST() and STD() are for
debugging output; S() and SD() always print.

Signed-off-by: Russell Steicke <russells@adelie.cx>