Allow the time to be set with a command.
The format of the time is very specific: "set hh:mm:ss A".

"set" can be upper or lower case, the time must contain six digits and
two colons on the order displayed above, and the trailing letter can
be "A" for AM or "P" for PM, and must be upper case.  Only one space
is allowed between "set" and the time, and between the time and "A" or
"P", and leading or trailing space is not allowed.

But it works.

Signed-off-by: Russell Steicke <russells@adelie.cx>